chore(heureka): renames view names
